agentes-csv-analyzer/requirements.txt
streamlit==1.32.0
pandas==2.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2
zipfile36==0.1.3
plotly==5.18.0 